Split mock api tests into multiple files and add comment describing test generation script
